# Conflicts:
#	Wave/Assets/Scripts/Wave/Wave.cs

+++ Wave/Assets/Scripts/Building/Building.cs
@@ -39,7 +39,7 @@ public abstract class Building : MonoBehaviour {
     }
     public float GetHP()
     {
-        return hp;
+        return (int)hp;
     }
     public float GetMaxHP()
     {
